计算人工平均薪酬

+++ licai/src/main/java/com/sjzx/entity/ImportantTarget.java
@@ -1,13 +1,14 @@
 package com.sjzx.entity;
 
 import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.extension.activerecord.Model;
-import java.util.Date;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.io.Serializable;
+import com.baomidou.mybatisplus.extension.activerecord.Model;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.Accessors;
+
+import java.io.Serializable;
+import java.util.Date;
 
 /**
  * <p>
@@ -85,7 +86,7 @@ public class ImportantTarget extends Model<ImportantTarget> {
     /**
      * 指标表t9
      */
-    private Integer t9;
+    private Long t9;
 
     /**
      * 备注
